Updated Azure Container App resource manager API Fixed build issue with Batch deployments in .NET 6 including upload and command options Updated AKS creation script to fix VNET assignment Updated Azure SDK packages to latest versions (requiring some code changes)

+++ src/SqlBuildManager.Console/Aci/TemplateClass.cs
@@ -28,7 +28,7 @@ namespace SqlBuildManager.Console.Aci
         public int Cpu { get; set; }
 
         [JsonPropertyName("memoryInGB")]
-        public int MemoryInGB { get; set; }
+        public double MemoryInGB { get; set; }
     }
 
     public class Resource
